+++ applications/main/bad_usb/helpers/ducky_script.h
@@ -15,6 +15,7 @@ typedef enum {
     BadUsbStateRunning,
     BadUsbStateDelay,
     BadUsbStateStringDelay,
+    BadUsbStateWaitForBtn,
     BadUsbStateDone,
     BadUsbStateScriptError,
     BadUsbStateFileError,