+++ applications/main/subghz/helpers/subghz_custom_event.h
@@ -13,6 +13,8 @@ typedef enum {
     SubmenuIndexSomfyTelis,
     SubmenuIndexBeninca433,
     SubmenuIndexBeninca868,
+    SubmenuIndexAllmatic433,
+    SubmenuIndexAllmatic868,
     SubmenuIndexIronLogic,
     SubmenuIndexElmesElectronic,
     SubmenuIndexSommer_FM_434,