+++ lib/subghz/protocols/keeloq.c
@@ -232,9 +232,9 @@ static bool subghz_protocol_keeloq_gen_data(
             decrypt = btn << 28 | (instance->generic.serial & 0xFF) << 16 | instance->generic.cnt;
         }
 
-        // Beninca -> 4bit serial - simple XOR
+        // Beninca / Allmatic -> no serial - simple XOR
         if(strcmp(instance->manufacture_name, "Beninca") == 0) {
-            decrypt = btn << 28 | (instance->generic.serial & 0xF) << 16 | instance->generic.cnt;
+            decrypt = btn << 28 | (0x000) << 16 | instance->generic.cnt;
         }
 
         if(strcmp(instance->manufacture_name, "Unknown") == 0) {